refactor(habitat): split diagnostic and host models Extract shared diagnostic contracts and host/protected-surface policy from the check module into named service-model domains. Grit provider code now imports diagnostic and host language directly instead of depending on check module internals, while baseline/source/structural check execution remains owned by the check module.
Validation run before commit:
- bun run --cwd tools/habitat-harness check
- bun nx run @internal/habitat-harness:boundaries
- bun run --cwd tools/habitat-harness validate:boundary-taxonomy
- bun run --cwd tools/habitat-harness validate:service-module-shape
- bun run --cwd tools/habitat-harness validate:grit-patterns
- focused habitat-harness tests: 123 passed
- provider/shared-domain scan: no check-module model imports
- router import scan: no non-module imports
- git diff --check
