+++ tools/habitat-harness/src/providers/grit/output.ts
@@ -13,7 +13,7 @@ import {
   type NativeGritCheckRequest,
   NativeGritCheckRequestSchema,
   nativeGritCheckRequestFromCommandResult,
-} from "@internal/habitat-harness/service/modules/check/model/index";
+} from "@internal/habitat-harness/service/model/diagnostics/index";
 import { type Static, Type } from "typebox";
 import { Value } from "typebox/value";
 import {
