+++ tools/habitat-harness/src/service/model/diagnostics/dto/diagnostic-catalog.schema.ts
