+++ tools/habitat-harness/src/service/model/diagnostics/dto/diagnostic-command.schema.ts
