+++ tools/habitat-harness/src/service/model/diagnostics/dto/diagnostic-identity.schema.ts
