+++ tools/habitat-harness/src/service/model/diagnostics/dto/diagnostic-outcome.schema.ts
