+++ tools/habitat-harness/src/service/model/diagnostics/errors/diagnostic-provider.errors.ts
