+++ tools/habitat-harness/src/service/model/host/policy/protected-zone-diagnostics.policy.ts
@@ -1,5 +1,5 @@
+import type { HabitatDiagnostic } from "@internal/habitat-harness/service/model/diagnostics/index";
 import type { RuleFileLayerFacts } from "@internal/habitat-harness/service/model/rules/index";
-import type { HabitatDiagnostic } from "@internal/habitat-harness/service/modules/check/model/policy/structural/schema";
 import type {
   DeclarationReadiness,
   ProtectedMutationDecision,
