+++ tools/habitat-harness/src/service/model/host/policy/protected-zone-guard.policy.ts
