+++ tools/habitat-harness/src/service/model/host/policy/protected-zone-path-actions.policy.ts
