+++ tools/habitat-harness/src/service/model/host/policy/protected-zone-scan-root.policy.ts
