+++ tools/habitat-harness/src/service/modules/check/model/policy/structural/schema.ts
@@ -1,21 +1,13 @@
+import {
+  type HabitatDiagnostic,
+  HabitatDiagnosticSchema,
+  type HabitatSeverity,
+} from "@internal/habitat-harness/service/model/diagnostics/index";
 import { type Static, Type } from "typebox";
 import { Value } from "typebox/value";
 
-export const HabitatSeveritySchema = Type.Union([Type.Literal("error"), Type.Literal("advisory")]);
-export type HabitatSeverity = Static<typeof HabitatSeveritySchema>;
-
-export const HabitatDiagnosticSchema = Type.Object(
-  {
-    ruleId: Type.String({ minLength: 1 }),
-    path: Type.String({ minLength: 1 }),
-    line: Type.Optional(Type.Number()),
-    message: Type.String({ minLength: 1 }),
-    severity: HabitatSeveritySchema,
-    baselined: Type.Boolean(),
-  },
-  { additionalProperties: false }
-);
-export type HabitatDiagnostic = Static<typeof HabitatDiagnosticSchema>;
+export type { HabitatDiagnostic, HabitatSeverity };
+export { HabitatDiagnosticSchema };
 
 export const RuleStatusSchema = Type.Union([
   Type.Literal("pass"),
